+++ packages/security/src/index.js
@@ -6,6 +6,3 @@
  */
 
 export { ComboButton, ComboButtonItem } from './ComboButton';
-export { Decorator } from './Decorator';
-export { ErrorPage } from './ErrorPage';
-export { ICA } from './ICA';