+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-core@0.5.1...@carbon/ibm-cloud-cognitive-core@0.5.2) (2021-01-26)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive-core
+
+
+
+
+
 ## [0.5.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-core@0.5.0...@carbon/ibm-cloud-cognitive-core@0.5.1) (2021-01-20)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive-core